fix waterForm and add WaterModal
