+++ src/components/modals/WaterModal/WaterModal.module.css
@@ -1,0 +1,69 @@
+.modal {
+  max-width: 343px;
+}
+
+.content {
+  padding-top: 20px;
+  padding-bottom: 20px;
+}
+
+.icon {
+  position: absolute;
+  top: 20px;
+  right: 20px;
+  stroke: var(--dark);
+  fill: transparent;
+}
+
+.title {
+  font-weight: 700;
+  font-size: 20px;
+  line-height: 1.2;
+  letter-spacing: -0.01em;
+  margin-bottom: 24px;
+  color: var(--dark-blue);
+}
+
+.choose {
+  font-weight: 700;
+  font-size: 16px;
+  line-height: 1.25;
+  letter-spacing: -0.01em;
+  color: var(--dark-blue);
+  margin-bottom: 20px;
+}
+.amount {
+  font-weight: 400;
+  font-size: 14px;
+  line-height: 1.28;
+  letter-spacing: -0.01em;
+  color: var(--dark);
+  margin-bottom: 8px;
+}
+
+.counter {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  margin-bottom: 16px;
+}
+
+.btnCounter {
+  border: 2px solid var(--dark-blue);
+  width: 40px;
+  height: 40px;
+  padding: 0;
+  border-radius: 50%;
+  font-weight: 700;
+}
+
+.countValue {
+  font-weight: 700;
+  font-size: 14px;
+  line-height: 1.28571;
+  text-align: center;
+  border-radius: 30px;
+  padding: 11px 20px;
+  color: var(--white);
+  background-color: var(--dark-blue);
+}
